20.06.25

+++ PrintF-v2/print_pointer.c
@@ -6,11 +6,11 @@
 /*   By: fekandle <fekandle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/20 16:17:20 by fekandle          #+#    #+#             */
-/*   Updated: 2025/06/20 17:46:43 by fekandle         ###   ########.fr       */
+/*   Updated: 2025/06/20 18:02:42 by fekandle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
 int	print_ptr(void *p)
 {
@@ -18,7 +18,7 @@ int	print_ptr(void *p)
 	int				count;
 
 	count = 0;
-	ptr = p;
+	ptr = (unsigned long)p;
 	count += print_str("0x");
 	count += unsign(ptr, 16, "0123456789abcdef", 1);
 	return (count);
